+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This page reports the main releases only and the main changes therein.
 ### Fixed
 - fixed threshold parameter behavior un unsharp mask filter
 - prevent failure in case a single file is selected for parallel alignment
+- reverts to sequential pyramid focus stacking if temporary disk space is insufficient
+
+### Changed
+- optimized temporary disk cache for parallel pyramid focus stacking
 
 --------------------------